image-upload-server/requirements.txt
Flask
google-cloud-vision
gunicorn  # Recommended for production deployments like Cloud Run